Mark Product Family form test as incomplete
It's getting really heavy to add the form type entity to the test form
factory. It wouldd require to mock many dependencies (especially
ManagerRegistry).

I'd like to open a discussion on the need of testing
such classes (FormType) as it, really often, means testing symfony
itself, instead of our added-value code.

I'd like to add that this part of the application will be tested by
functionnal tests. Indeed, if something is wrong in the form type that
is implemented, the form won't be rendered properly, thus (if made
right) making functionnal tests fail.