Refactored product fetching
Added a find method to the product manager
which will ensure that all values are present in the product (even the
ones that has not been persisted yet, like a new channel value).
Also refactored the product edit template by transforming the flat
collection of values into a tree of metadata and FormView.