Prevent encoding in csv array with non string keys
Because of the column normalization (enabling to encode data with
different number of attributes), encoding array with integer key must
used an other strategy than array_merge (in fact it must use union,
which changes the order of the attributes). So, to ease the process, and
because we expect to receive hashes with string keys in 99% of cases,
encoding array with integer keys has been forbidden.