Removed category form type usage from the category import
The category form type requires a security context token to display the
locale the current user have the right to edit. Unfortunately, the
import is run through command line, thus no token is available.

Anyway, the functionnal rule is to allow anyone who is launching the
script to import any data (meaning any locales). That is because import
will be most likely executed by cron script.