PIM-2205 Prevent product values from being unintentionnaly removed
Because the product edit form only contains values in a given language,
the odm adapter removes values from db that are not present in the collection.
This does not happen when using the orm.

Anyway, it makes sense to prevent the form from adding/removing values
as it will never happen currently (values are created on server side, so
extra value fields should be safely ignored).