Archiving big files causes huge memory usage
The ArchivableFileWriterArchiver is using streams to write to the zip
archive, but it reads the content of the source file using
file_get_contents.

This behaviour is changed to also use a stream to read the source file
and use putStream of flysystem to add content to the target archive.