Adding some tests for the search cases
At this point, the PM & P are indexed within the same index & type.
There is nothing done specially in the mapping to index the PM as they
have the same format as the P.

All the simple search use cases work because they only require to
retrieve the entity (PM or P) that actually holds the value searched.

The first use case failing is when we attempt to search for products
which values is held by the parent product model.

At this point, we also know that the _parent propertie can't reference a
document of the same type, nor can it reference a document from another
index.