PIM-6958: Add  sets of factories to load from DB
Since the 2.0, we introduced a set of value factories responsible for
instanciating product values depending on the attribute type. There is
currently ~ one value factory per attribute type.

Those factories are used by the ValueCollectionFactory which is then
capable of loading a list of objects given the raw data coming from DB (json
serialized representation of all the values for a product).

Keep in mind that those small factories are used for loading data from
DB and also to instanciate product values in the import jobs.

Those small factories, would each check the incoming data type (eg, check if it's
a string an array...) and when the data was related to some external
entities such as a file or option code, it would also check it's
existence using the corresponding repositories.

The issue appears when one of those linked entities would not be
available (eg, file deleted, or option code removed). Until now, the
factories would throw an exception which would be catched by the
ValueCollectionFactory (depending on the exception type, meaning, not
always). The thrown exception would make the ValueCollectionFactory skip
totally the product value, incapable to load anything if that was
possible.

Such a strict behavior is ok for import jobs: we want to give the most
precise feedback to a user and be quite defensive in what kind of data
we let in the PIM.

However, this behavior is way too strict in the context of loading the
product values from DB to display in the PEF. For instance, If a user
has selected options "A", "B", "C", "D" and "E" in the past and another
user has removed attribute code "E" from the attribute multiselect.

The next time a user loads the product, he will get a completely empty
attribute.

The solution I suggest here, is too recognize that we have two behaviors
for two different context and add more factories to cover all the cases:
  Strict for imports, gracefull to edit in the PEF.