Add specific reader for the compute desc job
When importing product models, 3 consecutive steps are executed
- Import root product models lines
- Import sub product models lines
- Compute product model descendants

At the end of the 2nd step, every sub & root product models are imported
in database (created or updated).
But they are not indexed, nor their product children' completeness is
 calculated / data reindexed.

The import of a file containing 500 root product models with 3 sub
product models (2000 lines in the file) has the following stats:
- Import root product models < 1 min
- Import sub product models ~ 1 min
*- Compute product model descendants 1h*

In all tests made, we observe the memory usage is stable (Thanks
@philippe Mossière).

Clearly, what we ought to do is optimize the 3rd step: compute product
model descendants.

Today this step is quite naïve:
for each product model that is found in the file to import we reindex
the whole "self and subtree" (the product model present in the file +
all its children) and also compute the completeness for products children.

**Dataset**:
Dataset 1:
- 1 file of product models containing: (2000 lines)
  - 500 root product models
  - 3 sub product models for each root product models (1500 sub product
  models)
- 1 file of products containing 1 product per sub product model (2000 lines)

**Solution implemented in this branch**:

Product models are organized in trees, and it's possible that one or
more children are present in the file to import. Hence, the compute
product model descendants job would run:
- Once for the root product model
- X more times for each children present in the file

While in reality, what we want is to run the job in the topmost product
model of the tree if possible.

So in this PR, what we do is prior to run the computing of the product
descendants for a sub product model, we check if his parent is already
present in the file.
If its the case, we skip this sub product model as the computing of his
parent will also run the computing for this sub product model.

We tried to import different files:
- Best case scenario: all the root product models are at the top of the
file
- Worst case scenario: all the root product models are at the end of the
file

Results:
Best case - Compute product model descendants execution time: 25 mins
Worst case - Compute product model descendants execution time: 25 mins

It seems like a substencial progression, but there are some drawbacks:
- this measures should decrease for bigger files let's say 15 or 20K Lines
to import.
- This fix becomes an overhead for files having only sub product models
to import (as we check for their parent's presence in the file)