CI - Fix issue "Warning: touch(): Unable to create"
The CI uses an extension provided by luigio/fastest to show all the
scenarios of the project.

This extension uses a formatter to show the message which is configured
in the '.ci/behat.community.yml' file with an output_path.

When a path that does not exist is set to this output_path, the
extension cannot create the file to output in resulting in this error.

Fix:
Make sure the path exists prior to list scenarios.