PIM-7132: Revert solution found for 1.6-1.7
@julien has reworked the way the mass-actions are created from the front
to the back.
The router component has changed as well as the mass-edit endpoint.

The solution found in 1.6/1.7 does not apply anymore.

Do we need to find a suitable solution ?

Maybe, the problem in 1.6 and 1.7 would occur when a user would select
~10 items in the datagrid, indeed the item would be passed to the
back-end through their Ids and we know how long a mongodbId is (24
chars).

In 2.0, there are no mongoIds anymore with the single storage and
firefox and chrome are removing the url length limit in their newer
versions.
There is still an apache limit (which is around 8K chars), but I think
it is safe to say that there is a few chances that a user will have this
issue in 2.0.

So for now, we will just revert the code made in 1.6 and 1.7 to fix this
issue.