PIM-7221: Fix upload assets
Prior to this fix:
Whenever an asset is uploaded all the products having an attribute of
type asset_collection containing the asset uploaded should have its
completeness reset.

In Mysql it's quite easy as all the work is handled in the DB, but in
mongodb a long mongo request was generated to unset the completeness of
all the documents having the asset.

This caused the PIM to freeze and eventually not terminate.

Solution:
This fix takes on a new perspective of resetting the completeness in
mongo. Instead of relying on a single request to reset the completeness,
we use the PQB to find the products which completeness need to be reset
and reset it.

By using the pqb, we also take advantage of the ElasticsearchBundle when
it is activated.