PIM-7347: Fix Completeness Calculation with locale specifics attribute
**Context:**

The calculation of the completeness of products is a two step process.

It basically starts in this class: `Pim\Component\Catalog\Completeness\CompletenessCalculator::Calculate`

1. Build the list of required values based on the attribute requirements.

Based on the attribute requirements of the family, we generate a
`Pim\Component\Catalog\EntityWithFamily\RequiredValueCollection`, which holds a list of `\Pim\Component\Catalog\EntityWithFamily\RequiredValue`.

Basically, this object holds a reference for each required value for:
- An attribute
- A channel
- A locale

Important note:

Conviniently, this RequiredValue extends `Pim\Component\Catalog\Model\ValueInterface` which
gives us the opportunity to get the value of a product based on this
object with
```
$productValueCollection->getSame($requiredValue); // the actual value
```

2. Use the required values and the values of the product to generate
completeness.

In
`Pim\Component\Catalog\EntityWithFamily\IncompleteValueCollectionFactory`, we use
this list of RequiredValues to find the value for a given product using
the *ValueCollection::getSame method* and test if it's complete or not.

And that's basically it!

**Problem: Let there be locale specific attributes**

The coupling between a required value (the information it holds) and the
fact that we can directly use this object to retrieve the data from the
product value collection becomes an issue when the locale specific comes
in.

Let's say we have:
- An attribute "description" which is localeSpecific on
'en_US' and 'fr_FR'.
- A family "camcorders" with two attributes: "sku", "description"
required on all channels
- A product "webcam" with the following values:
```
| sku | description |
| 001 | ""          |
```

Now, when we generate the required values in the first step for this family we get
something like this:
```
[
  'requiredValue1' => [
      'attribute' => 'description',
      'locale' => 'fr_FR',
      'scope' => null
  ],
  'requiredValue2' => [
      'attribute' => 'description',
      'locale' => 'en_US',
      'scope' => null
  ]
]
```

In the second step, we try to check if the values of the products are
complete or not, we try to get the value using:

```
$myWebcapProduct->getSame($requiredValue1); // => returns null
$myWebcapProduct->getSame($requiredValue2); // => returns null
```

indeed, the value collection cannot return it's data for the attribute
description on the locale 'fr_FR' or 'en_US' because it is not scopable!

Internally, in the value collection of the product they are indexed
using the keys:
'description-<all_channels>-<all_locales>' and not 'description-<all_channels>-fr_FR' nor
'description-<all_channels>-en_US'.

That's why we need to decouple the required value structure and the way
the value of the products are retrieved.

**Solution:**

When building the requiredValue make a distinction between the actual
requirements and the way we want to retrieve the product value.

A good way to access a value within a valueCollection is to use it's
keys (as shown above).

So we added a new property to the requiredValue class called:
"valueKey".

For locale specific attributes a required value will look like so:
```
[
  'requiredValue1' => [
      'attribute' => 'attribute_code',
      'locale' => 'locale_code',
      'scope' => null
      'valueKey' => 'attribute_code-<all_channels>-<all_locales>'
  ],
]
```

for all other attributes:
```
[
  'requiredValue1' => [
      'attribute' => 'attribute_code',
      'locale' => 'locale_code',
      'scope' => null
      'valueKey' => 'attribute_code-<all_channels>-<locale_code>'
  ],
]
```