PIM-6784: Fix identifier filter on proposal grid
**Before:**

The "SKU" filter on the proposal grid was used to search impacted
products with a proposal.

**Now:**

- The "SKU" filter on the proposal grid is used to search on
proposals with a change on a sku. It searches on the new value.

- A new filter has been introduced "Identifier" that is used to search
impacted products by a proposal.

**How:**

1. Do not show the "SKU" filter by default
Functionnaly:
The proposal datagrid was using the same configurator filter as the
product grid which shows the "SKU" filter by default.

Since it is not working the same on the proposal grid and the product
grid, we needed to treat it as any other filter on attributes in the
proposal grid and they are hidden by default (accessible from the
filter list)

Technically it means:
I imported the FilterConfigurator of the CE and put a copy of it in
the WorkflowBundle and deactivate the fact that the attribute on
identifier is shown by default.

Then changed the proposal grid before listener definition to use this
new configuratorFilter.

2. Creating the new Identifier filter

It's simply a new "filter" in the datagrid definition
(In PimEnterprise/Bundle/WorkflowBundle/Resources/config/datagrid/proposal.yml) that filters on the property "product_identifier".