PIM-7492: Add extensible factories to create attributes
*Pattern abstract factory*
*Pattern registry*

We figured we will add new attribute types available for the enriched
entities, so we designed an extension point so that it's easier for us
in the future.
As well as easier for the partners to add custom ones (we think this
extension point will be heavily needed).

There are multiple types of attributes, for now we have: 'TextAttribute'
and 'ImageAttribute'.

For each of those attributes is associated a factory responsible for
creating objects of those attribute types ('TextAttributeFactory',
'ImageAttributeFactory').

Those factories are registered into a registry.

The handler gets a command of type 'AbstractCreateAttributeCommand',
depending on it's concrete type 'CreateTextAttributeCommand' or
'CreateImageAttributeCommand', the registry will know which factory to
return.

Given the command, the handler gets the right factory from the
registry, creates the attribute object and adds it to the repository.