PIM-7492: Add CreateAttribute handler & tests
This commit adds the CreateAttributeHandler in the application layer
with spec and acceptance tests.

The create handler uses the attribute factory registry to
create the attribute given a command.

Before hand, it checks using query function the state of the
persistence.
- It checks the attribute it wants to create is unique given the
attribute identifier and the enriched entity identifier with
"ExistsAttribute::withIdentifier" function.
- It checks the order is not already taken by another attribute for the
enriched entity.

If one of this case is true, a \RuntimeException is thrown by the
handler and nothing is created.