PIM-7492: Add SqlAttributeRepository
Every common information to the attributes are set on one column in the
attribute table. Other additionnal information specific to concrete
types (such as text or image) are stored in Json in the "data" column.

There is a constraint on the composite key (identifier,
enriched_entity_identifier).
There is no DELETE CASCADE relationship with enriched entities.

For now, the attribute repository is not extensible.