PIM-7548: fix non injected serializer due to the split of normaliers
The service is sometimes directly injected into controllers, without injecting the serializer. As the serializer is not called, the service does not have the serializer loaded int to it, resulting in an error. It was working before because, somehow, the global serializer of the PIM was called before the call of this service. So, the serializer was always injected. It's not the case anymore by splitting serializers by format