PIM-7509: Calculate completeness of variant products on rule execution
**Context:**

In the PIM, variant products are directly updated on unitary save of
their parent.
Meaning if you use the UI to update a product model all his children
will be udpated (to calculate the completeness, and to reindex).

This is not the case on bulk save (due to performance mainly). So for
imports of product models, the children of the updated product models
are updated in a separate step, and only on this step.

**Problem:**

There is no concept of "step" whatsoever in the context of rules
execution. This means that until today, the children weren't updated at
all if one of their parent was updated by a rule.

**Proposed Solution:**

As the rules are already executed in an asynchronous context, we can
directly update the descendants of all the product models
impacted/updated by a rule right after the save of the product models.

The logic is the same as what's done in the 'Pim\Component\Catalog\Job\ComputeProductModelsDescendantsTasklet'
with the use of the productModelDescendantsSaver.