PIM-7509: Fix search on categories
**Context:**

Following the wonderful work made by by @ahocquard on improving the
scalability of the search on categories, the "aggregated" search of
product models with categories was not working properly.

**Problem:**

In his refactoring, he introduces a new way to filter on categories on
the product grid. Before his changes when a user would select a category
in the category tree with the "include children" flag set to "yes". The
whole list of categories would be passed on to the back-end (category
selected + children) instead of using the IN_CHILDREN operator available
with the category filter.

@ahocquard fixes does just that, it uses the IN_CHILDREN operator
instead of the IN_LIST with a big list of category codes.

**Solution:**

Make sure the PQB of products and product models in enrich takes this new
case into account and aggregates the result when a filter with
"IN_CHILDREN" is used.