Fix failing test when deleting a root product model
When loading the fixtures with the test container and then deleting the root product model with the remover service from the test container, the delete cascade is not applied. The root cause is not determined. The solution to make it works is to not use the remover service from the test container.