PIM-8605: Fix validation message (show asset family identifier instead of label)
Two things in this commit:

1. Rename the CreateAssetFamilyCommand->$code property

Because the has EditAssetFamilyCommand->$identifier, I made this choice
to have the same property on both commands, so when a validator is
applied to both commands, it only has to access $identifier instead of
using the instanceof operator to guess which property to access to get
the identifier of the asset family.

(I chose $identifier instead of $code because that's how the domain
model is made).

I also renamed some validation services laying around that would
reference the "code" instead of the identifier

2. Fixing violation messages

In some violation messages such as
(src/Akeneo/AssetManager/back/Infrastructure/Validation/AssetFamily/ThereShouldBeLessAssetFamilyThanLimitValidator.php)
we would put the label of the asset family instead of its identifier.

Why this change? because if the asset family doesn't have any label, the
violation message would print an empty string ("") to identify the asset
family.