PIM-8605: Check the channel/locale even if there are some extrapolations
Before this commit, when there was an extrapolation in a product
selection or an assignement, we would not validate anything because we
thought it would be hard to partially validate those objects.

Talking about it later on, we figured we could still validate the
channels and locales when they are static (eg, "ecommerce", "fr_FR").