PIM-8605: Prevent crashing when inserting an invalid extrapolated attribute code
When inserting a product link rule with an invalid extrapolated attribute code
such as "my attribute code is invalid because it has spaces", the
PIM would crash because later on, we use a query function to check if
the attribute actually exists using a query function that uses domain
value objects.

Those objects cannot be created hence an exception is thrown. The
approach I chose is to reuse the Attribute\Code.php constraint
validation to check if the attribute code is valid before using the
query function.

(another approach could have been to introduce a new query function that
takes strings as arguments and return true/false independently if the
attribute code is valid or if it does not exist).

ERRATUM:

Reusing the whole Attribute\Code validation constraint ended up in not
being a good idea since it also checks for things we actually want:
- This validation checks that the attribute code cannot be "code",
"image" or labels" BUT we want the user to be able those in an
extrapolated attribute.

So I ended up, copy/pasting only the few constraint I needed to check
the attribute code was a valid one.