PIM-8781: Hide assets analytics from the system info panel
By moving the service declaration of the classes related to the analytics from AssetManager into Enrichment (where they do belong
by the way). We take advantage of the feature flag made by KriKri, hence
the analytics do not show up in the system info panel, when the
AkeneoAssetManagerEnrichmentBundle is not activated.