PIM-8779: Fix execution of "replace" product link rules
Issue:
Product link rules having a "replace" action won't be executed
correctly.

Fix:
In the rule engine, Add action rely on an "items" property of the action
(contained in the rule definition) to set this value onto the product
value, wherehas "Set" action rely on a
"value" property of the action.

For this fix, I suggest we always build a rule definition having an
"items" and a "value" key independently of the type of action we are
doing. Hence, the rule engine will always be able to execute the rule
definition.