PIM-8779: Fix when setting an unknown product attribute in the product link rule
Issue:
When trying to set a product attribute that does not exist in the
product link rule assignment result in a 500 error.

Diagnostic:
This would happen because, even if we validate if the product attribute
exist or not thanks to the validator, we would still try to check if the
product attribute points to the asset family we are setting the product
link rule for.

Fix:
We only check if the product attribute has the right asset family if the
rule engine validator does not return any violations.