TIP-1073: declare service as lazy as a workaround
After a lot of investigation, we choosed to do this workaround for now. The container is injecting the wrong unit of work in the locale reposiory. Therefore, when you try to persist a locale by saving it in the Locale Saver, it's another UOW that is used. As this unit of work is different, the locale is considered as NEW by the UOW when trying to flush it. So, it tries to INSERT it instead of updating it: we have a conflict in database on the primary key.