mitigate deadlock problem on quality insight
The mass insertion in quality insight table
'pimee_data_quality_insights_criteria_evaluation' was causing dead locks
in MySQL. This patch makes the application to retry up to 4 times with a
random sleep time between each retry. If it does not work, then the
table is locked and the insertion is forced.