fix product table lock due to foreign key
When importing lot of data with parallel batches, the completeness table
could experience deadlocks. To prevent this a fail & retry mechanism has
been set up. For the last retry the completeness table is entirely
locked to allow the process to write data in it. Due to the foreign key,
this also locked the product table which in some cases lead to soft
dead-lock situation hanging the whole database.
This patch removes the foreign key, treating this tables as a
denormalization table which is NOT the source of truth regarding the
completeness.