PIM-9611: force the order of a join due to statistics bias for low cardinality
The number of extimated cardinality is very far from the reality for completeness table. This is because the number of channel is a very low cardinality. See https://www.percona.com/blog/2017/09/11/updating-innodb-table-statistics-manually/. When doing a join between the completeness , products and channel tables, in some cases, the cost based optimizer thinks that the channel is a better candidate than products for filtering. Actually it does not filter anything and it reads the whole completeness table index, and then it filters the products. If the completeness table has more than 10 millions lines, it filters the products over these 10 millions lines without any index. And it is very slow.