Product API: include search_locale in pagination links
Previously, the pagination links in both Product and ProductModel list
api responses would have the search_locale param stripped if it was
used. This meant that the first page would be retrieved, but then
following the nextLink url in the response would throw a 422 when
visited (because of a lack of search_locale).