fix #14117: check for cookies property in ScriptNonceGenerator (master)
When the command `bin/console debug:event-dispatcher` is used the `ScriptNonceGenerator::getGeneratedNonce` is called. At this point the `ScriptNonceGenerator` object has been constructed with an `RequestStack` which has an empty Request object inside missing the `cookies` property.

The fix checkes whether a request was found (a null pointer can occure based on the `getCurrentRequest` method signature) and if the `cookies` property is present before using it to fill the `bapId` variable. If those conditions are not met, an empty string is used as a placeholder.

Github issue for Akeneo master https://github.com/akeneo/pim-community-dev/issues/14117