Apply sorting in the versioned normalizer for compare
Hi,

We have a case with a small catalog, when importing full file date, the version manager reports that changes have been made every time a product was imported, even if the product is the same. Creating unwanted version history changes. 

This change helps to mitigate that problem although maybe more a elegant solution might be possible.