Update ScalarValueFactory.php
For whatever reason, if a csv import with invalid information is supplied, this can sometimes create an invalid product that will crash when loaded, can't be deleted, and stops syncs. I could not debug further, but for some reason localeCode is null.
Checking if the localeCode is null lets everything load.