Fix invalid JSON Path Expression
A JSON Path Expression in the form `$.memberName` is only valid if
`memberName` is a field name that does not contain spaces, dots and/or
other characters or starts with a number!
So if there is an AttributeCode that starts with a number, e.g.
`1ws_thumbnail`, this leads to an invalid JSON Path Expression.
This can be avoided with this notation `$. "memberName"`.