fix: do not insert already existing jobs in GE
The job clean_table_values_following_deleted_options is not always here in GE, because it is created on the fly if needed. It should always exist. Therefore, when inserting missing job instance from EE in GE migration, if the job clean_table_values_following_deleted_options already exist, it does not fail anymore. If it does not exist yet, it will create it.